+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## main / unreleased
 
+## 3.7.2 / 2025-10-22
+
+* [BUGFIX] AWS SD: Fix AWS SDK v2 credentials handling for EC2 and Lightsail discovery. #17355
+* [BUGFIX] AWS SD: Load AWS region from IMDS when not set. #17376
+* [BUGFIX] Relabeling: Fix `labelmap` action validation with the legacy metric name validation scheme. #17372
+* [BUGFIX] PromQL: Fix parsing failure when `anchored` and `smoothed` are used as metric names and label names. #17353
+* [BUGFIX] PromQL: Fix formatting of range vector selectors with `smoothed`/`anchored` modifier. #17354
+
 ## 3.7.1 / 2025-10-16
 
 * [BUGFIX] OTLP: Prefix `key_` to label name when translating an OTel attribute name starting with a single underscore, and keep multiple consecutive underscores in label name when translating an OTel attribute name. This reverts the breaking changes introduced in 3.7.0. #17344
@@ -48,7 +56,7 @@
 * [BUGFIX] TSDB: Avoid misleading `Failed to calculate size of \"wal\" dir` error logs during WAL clean-up. #17006
 * [BUGFIX] TSDB: Prevent erroneously dropping series records during WAL checkpoints. #17029
 * [BUGFIX] UI: Fix redirect to path of `-web.external-url` if `-web.route-prefix` is configured. #17240
-* [BUGIFX] Remote-write: Do not panic on invalid symbol table in remote-write 2.0. #17160
+* [BUGFIX] Remote-write: Do not panic on invalid symbol table in remote-write 2.0. #17160
 
 ## 3.6.0 / 2025-09-17